revert documentation changes
