+++ components/radio/radio.vue
@@ -1,8 +1,6 @@
 <template>
   <label>
-    <div
-      :class="['d-radio-group', { 'd-radio-group--disabled': internalDisabled }]"
-    >
+    <div :class="['d-radio-group', { 'd-radio-group--disabled': internalDisabled }]">
       <div class="d-radio__input">
         <input
           :checked="internalChecked"
